pyproject.toml
[project]
name = "pailotv2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "faiss-cpu>=1.13.2",
    "fastapi[standard]>=0.128.0",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.8",
    "langchain-google-genai>=4.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "mcp>=1.26.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "python-pptx>=1.0.2",
    "pyyaml>=6.0.3",
]
